requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
websockets==12.0
pyjwt==2.8.0
cryptography==41.0.7
python-dotenv==1.0.0
aiosqlite==0.19.0
